requirements.txt
flask
google-genai
waitress
logging
os
base64
